tracking_service/requirements.txt
fastapi
sqlalchemy
uvicorn
passlib[bcrypt]
pyjwt
psycopg2-binary
pika
httpx
loguru
alembic
